feat: add vero `vc_type` (#827)
Adds the Vero validator client as a supported `vc_type`.

The Vero VC doesn't handle keys and requires a remote signer to be used
(`use_remote_signer: true`). It can handle the minimal preset, as well
as custom `seconds_per_slot` values.

I've added the "Vero - Detailed" Grafana dashboard. I'm not sure if it's
a good fit for this repo, let me know if you'd like me to remove it.

---------

Co-authored-by: Barnabas Busa <barnabas.busa@ethereum.org>